LayoutTests/storage/storagemanager/nested-dedicated-workers.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
jsTestIsAsync = true;
worker = startWorker('resources/nested-dedicated-workers.js');
</script>
</body>
</html>
